+++ README.md
@@ -1,2 +1,4 @@
-# crapage
+# Zabbix Metric Shipper
 Make data exported by Zabbix fly!
+
+# Configurarion
