--- crapage.yaml
@@ -1,14 +1,0 @@
-server_config: ./zabbix_server.conf
-targets:
-  - name: stdout
-    type: print
-    connection: stderr
-  # - name: tables
-  #   type: azuretable
-  #   connection: connstring
-# exports:
-#   history:
-#     - tables
-#     - stdout
-#   trends:
-#     - tables
